% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_output_dates.R
\name{.generate_output_dates}
\alias{.generate_output_dates}
\title{Generate output dates for forecast method.
This is a private function of 'nixtlar'}
\usage{
.generate_output_dates(df_info, freq, h)
}
\arguments{
\item{df_info}{A data frame that is created by the forecast method with the last dates of every unique id.}

\item{freq}{The frequency of the data, as a period or offset alias.}

\item{h}{The forecast horizon.}
}
\value{
A data frame with dates for the forecast.
}
\description{
Generate output dates for forecast method.
This is a private function of 'nixtlar'
}
\examples{
\dontrun{
  dates_df <- .generate_output_dates(df_info, freq, h)
}

}
\keyword{internal}
