% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_client_setup.R
\name{.get_client_steup}
\alias{.get_client_steup}
\title{Get NIXTLA_API_KEY from options or from .Renviron
This is a private function of 'nixtlar'}
\usage{
.get_client_steup()
}
\value{
If available, the NIXTLA_API_KEY. Otherwise it returns an error and a message asking the user to set the 'API' key.
}
\description{
Get NIXTLA_API_KEY from options or from .Renviron
This is a private function of 'nixtlar'
}
\examples{
\dontrun{
  .get_api_key()
}

}
\keyword{internal}
