% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_client_version.R
\name{.get_client_version}
\alias{.get_client_version}
\title{Get the 'nixtlar' client version string
This is a private function of 'nixtlar'}
\usage{
.get_client_version()
}
\value{
A character string of the form \code{"R-x.y.z"}, where \code{x.y.z} is the
installed 'nixtlar' package version (read from the DESCRIPTION file).
}
\description{
Get the 'nixtlar' client version string
This is a private function of 'nixtlar'
}
\examples{
\dontrun{
  .get_client_version()
}

}
\keyword{internal}
