% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_model_params.R
\name{.get_model_params}
\alias{.get_model_params}
\title{Retrieve parameters for 'TimeGPT' model
This is a private function of 'nixtlar'}
\usage{
.get_model_params(model, freq)
}
\arguments{
\item{model}{Model to use, either "timegpt-1" or "timegpt-1-long-horizon".}

\item{freq}{Frequency of the data.}
}
\value{
A list with the model's input size and horizon
}
\description{
Retrieve parameters for 'TimeGPT' model
This is a private function of 'nixtlar'
}
\examples{
\dontrun{
  .get_model_params()
}

}
\keyword{internal}
