% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/level_from_quantiles.R
\name{.level_from_quantiles}
\alias{.level_from_quantiles}
\title{Obtain level from  quantiles
This is a private function of 'nixtlar'}
\usage{
.level_from_quantiles(quantiles)
}
\arguments{
\item{quantiles}{A vector with the quantiles.}
}
\value{
A list containing the level vector and a data frame with the quantiles and their corresponding levels.
}
\description{
Obtain level from  quantiles
This is a private function of 'nixtlar'
}
\examples{
.level_from_quantiles(c(0.1, 0.5, 0.9))

}
\keyword{internal}
