% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/partition_payload.R
\name{.partition_payload}
\alias{.partition_payload}
\title{Partition payload
This is a private function of 'nixtlar'}
\usage{
.partition_payload(payload, num_partitions)
}
\arguments{
\item{payload}{A list containing the complete dataset for 'TimeGPT'.}

\item{num_partitions}{A positive integer, "auto", or NULL specifying the number of partitions. When set to "auto", the number of partitions is equal to the number of available cores. When NULL, it defaults to a single partition.}
}
\value{
A list of payloads, each representing a partition.
}
\description{
Partition payload
This is a private function of 'nixtlar'
}
\examples{
\dontrun{
  payload_lists <- .partition_payload(timegpt_data, num_partitions)
}

}
\keyword{internal}
