% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform_output_dates.R
\name{.transform_output_dates}
\alias{.transform_output_dates}
\title{Transform and format dates in 'TimeGPT' output
This is a private function of 'nixtlar'}
\usage{
.transform_output_dates(fcst, freq, flag)
}
\arguments{
\item{fcst}{Dataframe with the 'TimeGPT' output, where column 'ds' contains date strings.}

\item{freq}{Frequency of the data.}

\item{flag}{Indicator where 1 denotes 'tsibble' and 0 denotes 'dataframe'.}
}
\value{
'TimeGPT' output with the formatted dates
}
\description{
Transform and format dates in 'TimeGPT' output
This is a private function of 'nixtlar'
}
\examples{
\dontrun{
  fcst <- .transform_output_dates(fcst, freq, flag)
}

}
\keyword{internal}
