% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{electricity}
\alias{electricity}
\title{Electricity dataset}
\format{
\subsection{\code{electricity}}{

A data frame with 8400 rows and 3 columns:
\describe{
\item{unique_id}{Unique identifiers of the electricity markets.}
\item{ds}{Date in format YYYY:MM:DD hh:mm:ss.}
\item{y}{Price for the given market and date.}
}
}
}
\source{
\url{https://raw.githubusercontent.com/Nixtla/transfer-learning-time-series/main/datasets/electricity-short.csv}
}
\usage{
electricity
}
\description{
Contains prices of different electricity markets.
}
\keyword{datasets}
