% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{electricity_future_exo_vars}
\alias{electricity_future_exo_vars}
\title{Future values for the electricity dataset with exogenous variables}
\format{
\subsection{\code{electricity_future_exo_vars}}{

A data frame with 120 rows and 11 columns:
\describe{
\item{unique_id}{Unique identifiers of the electricity markets.}
\item{ds}{Date in format YYYY:MM:DD hh:mm:ss.}
\item{Exogenous1}{An external factor influencing prices. For all markets, some form of day-ahead load forecast.}
\item{Exogenous2}{An external factor influencing prices.
For "BE" and "FR" markets, the day-ahead generation forecast.
For "NP", the day-ahead wind generation forecast.
For "PJM", the day-ahead load forecast in a specific zone.
For "DE", the aggregated day-ahead wind and solar generation forecasts.
}
\item{day_0}{Binary variable indicating weekday.}
\item{day_1}{Binary variable indicating weekday.}
\item{day_2}{Binary variable indicating weekday.}
\item{day_3}{Binary variable indicating weekday.}
\item{day_4}{Binary variable indicating weekday.}
\item{day_5}{Binary variable indicating weekday.}
\item{day_6}{Binary variable indicating weekday.}
}
}
}
\source{
\url{https://raw.githubusercontent.com/Nixtla/transfer-learning-time-series/main/datasets/electricity-short-future-ex-vars.csv}
}
\usage{
electricity_future_exo_vars
}
\description{
Contains the future values of the exogenous variables of the electricity dataset (24 steps-ahead). To be used with \code{electricity_exo_vars}.
}
\keyword{datasets}
