% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_frequency.R
\name{find_frequency}
\alias{find_frequency}
\title{Find frequency of time series data}
\usage{
find_frequency(df, time_col)
}
\arguments{
\item{df}{A tsibble or a data frame with time series data.}

\item{time_col}{Column that contains each time step.}
}
\value{
An alias (character) for the frequency.
}
\description{
Find frequency of time series data
}
