% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/infer_frequency.R
\name{infer_frequency}
\alias{infer_frequency}
\title{Infer frequency of a data frame.}
\usage{
infer_frequency(df, freq)
}
\arguments{
\item{df}{A data frame with time series data.}

\item{freq}{The frequency of the data as specified by the user; NULL otherwise.}
}
\value{
The inferred frequency.
}
\description{
Infer frequency of a data frame.
}
\examples{
df <- nixtlar::electricity
freq <- NULL
infer_frequency(df, freq)

}
