% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nixtla_client_forecast.R
\name{nixtla_client_forecast}
\alias{nixtla_client_forecast}
\title{Sequential version of 'nixtla_client_forecast'
This is a private function of 'nixtlar'}
\usage{
nixtla_client_forecast(
  df,
  h = 8,
  freq = NULL,
  id_col = "unique_id",
  time_col = "ds",
  target_col = "y",
  X_df = NULL,
  level = NULL,
  quantiles = NULL,
  finetune_steps = 0,
  finetune_loss = "default",
  clean_ex_first = TRUE,
  add_history = FALSE,
  model = "timegpt-1"
)
}
\arguments{
\item{df}{A data frame with time series data.}

\item{h}{Forecast horizon.}

\item{freq}{Frequency of the data.}

\item{id_col}{Column that identifies each series.}

\item{time_col}{Column that identifies each timestep.}

\item{target_col}{Column that contains the target variable.}

\item{X_df}{A tsibble or a data frame with future exogenous variables.}

\item{level}{The confidence levels (0-100) for the prediction intervals.}

\item{quantiles}{Quantiles to forecast. Should be between 0 and 1.}

\item{finetune_steps}{Number of steps used to finetune 'TimeGPT' in the new data.}

\item{finetune_loss}{Loss function to use for finetuning. Options are: "default", "mae", "mse", "rmse", "mape", and "smape".}

\item{clean_ex_first}{Clean exogenous signal before making the forecasts using 'TimeGPT'.}

\item{add_history}{Return fitted values of the model.}

\item{model}{Model to use, either "timegpt-1" or "timegpt-1-long-horizon". Use "timegpt-1-long-horizon" if you want to forecast more than one seasonal period given the frequency of the data.}
}
\value{
'TimeGPT''s forecast.
}
\description{
Sequential version of 'nixtla_client_forecast'
This is a private function of 'nixtlar'
}
\examples{
\dontrun{
  nixtlar::nixtla_set_api_key("YOUR_API_KEY")
  df <- nixtlar::electricity
  fcst <- nixtlar::nixtla_client_forecast(df, h=8, id_col="unique_id", level=c(80,95))
}

}
\keyword{internal}
