% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nixtla_set_token.R
\name{nixtla_set_token}
\alias{nixtla_set_token}
\title{Set token in global environment}
\usage{
nixtla_set_token(token)
}
\arguments{
\item{token}{The user's token. Get yours here: https://dashboard.nixtla.io/}
}
\value{
A message indicating the token has been set in the global environment.
}
\description{
Set token in global environment
}
