% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nixtla_validate_token.R
\name{nixtla_validate_token}
\alias{nixtla_validate_token}
\title{Validate token}
\usage{
nixtla_validate_token()
}
\value{
A message indicating whether the token is valid.
}
\description{
Validate token
}
\examples{
\dontrun{
  nixtlar::nixtla_set_token("YOUR_TOKEN")
  nixtlar::nixtla_validate_token
}

}
