% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_data.R
\name{prepare_data}
\alias{prepare_data}
\title{Prepare time series data for TimeGPT's API}
\usage{
prepare_data(df)
}
\arguments{
\item{df}{A tsibble or a data frame with time series data.}
}
\value{
A list with the time series data for TimeGPT's API.
}
\description{
Prepare time series data for TimeGPT's API
}
