% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_multi_series.R
\name{prepare_multi_series}
\alias{prepare_multi_series}
\title{Prepare multiple time series data for TimeGPT}
\usage{
prepare_multi_series(df, id_col, time_col, target_col)
}
\arguments{
\item{df}{A tsibble or a data frame with multiple time series data.}

\item{id_col}{Column that identifies each series.}

\item{time_col}{Column that identifies each timestep.}

\item{target_col}{Column that contains the target variable.}
}
\value{
A list with the time series data that will be sent to TimeGPT.
}
\description{
Prepare multiple time series data for TimeGPT
}
