% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_single_series.R
\name{prepare_single_series}
\alias{prepare_single_series}
\title{Prepare single time series data for TimeGPT}
\usage{
prepare_single_series(df, time_col, target_col)
}
\arguments{
\item{df}{A tsibble or a data frame with a single time series data.}

\item{time_col}{Column that identifies each timestep.}

\item{target_col}{Column that contains the target variable.}
}
\value{
A list with the time series data that will be sent to TimeGPT.
}
\description{
Prepare single time series data for TimeGPT
}
