% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_token.R
\name{set_token}
\alias{set_token}
\title{Set token in global environment}
\usage{
set_token(token)
}
\arguments{
\item{token}{The user's token. Get yours here: https://dashboard.nixtla.io/}
}
\value{
A message indicating the token has been set in the global environment.
}
\description{
Set token in global environment
}
