% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timegpt_anomaly_detection.R
\name{timegpt_anomaly_detection}
\alias{timegpt_anomaly_detection}
\title{Detect anomalies with TimeGPT}
\usage{
timegpt_anomaly_detection(
  df,
  freq = NULL,
  id_col = NULL,
  time_col = "ds",
  target_col = "y",
  X_df = NULL,
  level = c(99),
  clean_ex_first = TRUE
)
}
\arguments{
\item{df}{A tsibble or a data frame with time series data.}

\item{freq}{Frequency of the data.}

\item{id_col}{Column that identifies each series.}

\item{time_col}{Column that identifies each timestep.}

\item{target_col}{Column that contains the target variable.}

\item{X_df}{A tsibble or a data frame with future exogenous variables.}

\item{level}{The confidence level (0-100) for the prediction interval used in anomaly detection. Default is 99.}

\item{clean_ex_first}{Clean exogenous signal before making the forecasts using TimeGPT.}
}
\value{
A tsibble or a data frame with the anomalies detected in the historical period.
}
\description{
Detect anomalies with TimeGPT
}
