% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timegpt_data.R
\name{timegpt_data}
\alias{timegpt_data}
\title{Prepares data for TimeGPT's API}
\usage{
timegpt_data(df, freq, id_col, time_col, target_col)
}
\arguments{
\item{df}{A tsibble or a data frame with time series data.}

\item{freq}{Frequency of the data.}

\item{id_col}{Column that identifies each series.}

\item{time_col}{Column that identifies each timestep.}

\item{target_col}{Column that contains the target variable.}
}
\value{
A list with the given/inferred frequency and the prepared data.
}
\description{
Prepares data for TimeGPT's API
}
