% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timegpt_forecast.R
\name{timegpt_forecast}
\alias{timegpt_forecast}
\title{Generate TimeGPT forecast}
\usage{
timegpt_forecast(
  df,
  h = 8,
  freq = NULL,
  id_col = NULL,
  time_col = "ds",
  target_col = "y",
  X_df = NULL,
  level = NULL,
  finetune_steps = 0,
  clean_ex_first = TRUE,
  add_history = FALSE
)
}
\arguments{
\item{df}{A tsibble or a data frame with time series data.}

\item{h}{Forecast horizon.}

\item{freq}{Frequency of the data.}

\item{id_col}{Column that identifies each series.}

\item{time_col}{Column that identifies each timestep.}

\item{target_col}{Column that contains the target variable.}

\item{X_df}{A tsibble or a data frame with future exogenous variables.}

\item{level}{The confidence levels (0-100) for the prediction intervals.}

\item{finetune_steps}{Number of steps used to finetune TimeGPT in the new data.}

\item{clean_ex_first}{Clean exogenous signal before making the forecasts using TimeGPT.}

\item{add_history}{Return fitted values of the model.}
}
\value{
TimeGPT forecasts for point predictions and probabilistic predictions (if level is not NULL).
}
\description{
Generate TimeGPT forecast
}
