% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timegpt_historic.R
\name{timegpt_historic}
\alias{timegpt_historic}
\title{Generate TimeGPT forecast for the in-sample period (historical period).}
\usage{
timegpt_historic(
  df,
  freq = NULL,
  id_col = NULL,
  time_col = "ds",
  target_col = "y",
  level = NULL,
  finetune_steps = 0,
  clean_ex_first = TRUE
)
}
\arguments{
\item{df}{A tsibble or a data frame with time series data.}

\item{freq}{Frequency of the data.}

\item{id_col}{Column that identifies each series.}

\item{time_col}{Column that identifies each timestep.}

\item{target_col}{Column that contains the target variable.}

\item{level}{The confidence levels (0-100) for the prediction intervals.}

\item{finetune_steps}{Number of steps used to finetune TimeGPT in the new data.}

\item{clean_ex_first}{Clean exogenous signal before making the forecasts using TimeGPT.}
}
\value{
TimeGPT's forecast for the in-sample period.
}
\description{
Generate TimeGPT forecast for the in-sample period (historical period).
}
\examples{
\dontrun{
  nixtlar::nixtla_set_token("YOUR_TOKEN")
  df <- nixtlar::electricity
  fcst <- nixtlar::timegpt_historic(df, id_col="unique_id", level=c(80,95))
}

}
