% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timegpt_plot.R
\name{timegpt_plot}
\alias{timegpt_plot}
\title{Plot forecast and insample values.}
\usage{
timegpt_plot(
  df,
  fcst = NULL,
  h = NULL,
  id_col = NULL,
  time_col = "ds",
  target_col = "y",
  unique_ids = NULL,
  max_insample_length = NULL,
  plot_anomalies = FALSE
)
}
\arguments{
\item{df}{A tsibble or a data frame with time series data (insample values).}

\item{fcst}{A tsibble or a data frame with the TimeGPT point forecast and the prediction intervals (if available).}

\item{h}{Forecast horizon.}

\item{id_col}{Column that identifies each series.}

\item{time_col}{Column that identifies each timestep.}

\item{target_col}{Column that contains the target variable.}

\item{unique_ids}{Time series to plot. If NULL (default), selection will be random.}

\item{max_insample_length}{Max number of insample observations to be plotted.}

\item{plot_anomalies}{Whether or not to plot anomalies.}
}
\value{
Plot with the forecast and insample values
}
\description{
Plot forecast and insample values.
}
