% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_token.R
\name{.get_token}
\alias{.get_token}
\title{Get TIMEGPT_TOKEN from options or from .Renviron
This is a private function of nixtlar}
\usage{
.get_token()
}
\value{
If available, the TIMEGTP_TOKEN
}
\description{
Get TIMEGPT_TOKEN from options or from .Renviron
This is a private function of nixtlar
}
\keyword{internal}
