% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_request.R
\name{.make_request}
\alias{.make_request}
\title{Make requests to the 'TimeGPT' API
This is a private function of 'nixtlar'}
\usage{
.make_request(base_url, api_key, payload_list)
}
\arguments{
\item{base_url}{String specifying the API endpoint to which the request is sent.}

\item{api_key}{The user's API key.}

\item{payload_list}{List containing the information to be sent to the 'TimeGPT' API.}
}
\value{
List representing the JSON response from the API endpoint.
}
\description{
Make requests to the 'TimeGPT' API
This is a private function of 'nixtlar'
}
\examples{
\dontrun{
  response <- .make_request(base_url, api_key, payload_list)
}

}
\keyword{internal}
