% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_exogenous.R
\name{.validate_exogenous}
\alias{.validate_exogenous}
\title{Validate future exogenous variables (if applicable)
This is a private function of 'nixtlar'}
\usage{
.validate_exogenous(df, h, X_df)
}
\arguments{
\item{df}{A tsibble or a data frame with time series data.}

\item{h}{Forecast horizon.}

\item{X_df}{A tsibble or a data frame with future exogenous variables.}
}
\value{
Returns a vector with the exogenous variable names if the validation is successful. If the validation fails, it stops execution and returns an error message, indicating what went wrong.
}
\description{
Validate future exogenous variables (if applicable)
This is a private function of 'nixtlar'
}
\examples{
\dontrun{
df <- nixtlar::electricity_exo_vars
X_df <- nixtlar::electricity_future_exo_vars
.validate_exogenous(df, h=24, X_df)
}

}
\keyword{internal}
