% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nixtla_client_historic.R
\name{nixtla_client_historic}
\alias{nixtla_client_historic}
\title{\code{TimeGPT} historic forecast}
\usage{
nixtla_client_historic(
  df,
  freq = NULL,
  id_col = NULL,
  time_col = "ds",
  target_col = "y",
  level = NULL,
  quantiles = NULL,
  finetune_steps = 0,
  finetune_depth = 1,
  finetune_loss = "default",
  clean_ex_first = TRUE,
  model = "timegpt-1"
)
}
\arguments{
\item{df}{A tsibble or a data frame with time series data.}

\item{freq}{Frequency of the data.}

\item{id_col}{Column that identifies each series.}

\item{time_col}{Column that identifies each timestep.}

\item{target_col}{Column that contains the target variable.}

\item{level}{The confidence levels (0-100) for the prediction intervals.}

\item{quantiles}{Quantiles to forecast. Should be between 0 and 1.}

\item{finetune_steps}{Number of steps used to fine-tune 'TimeGPT' in the new data.}

\item{finetune_depth}{The depth of the fine-tuning. Uses a scale from 1 to 5, where 1 means little fine-tuning and 5 means that the entire model is fine-tuned.}

\item{finetune_loss}{Loss function to use for fine-tuning. Options are: "default", "mae", "mse", "rmse", "mape", and "smape".}

\item{clean_ex_first}{Clean exogenous signal before making the forecasts using 'TimeGPT'.}

\item{model}{Model to use, either "timegpt-1" or "timegpt-1-long-horizon". Use "timegpt-1-long-horizon" if you want to forecast more than one seasonal period given the frequency of the data.}
}
\value{
'TimeGPT''s forecast for the in-sample period.
}
\description{
\code{TimeGPT} historic forecast
}
\examples{
\dontrun{
  nixtlar::nixtla_set_api_key("YOUR_API_KEY")
  df <- nixtlar::electricity
  fcst <- nixtlar::nixtla_client_historic(df, id_col="unique_id", level=c(80,95))
}

}
\keyword{private}
