% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nixtla_client_setup.R
\name{nixtla_client_setup}
\alias{nixtla_client_setup}
\title{Set base 'ULR' and 'API' key in global environment}
\usage{
nixtla_client_setup(base_url = NULL, api_key = NULL)
}
\arguments{
\item{base_url}{Custom base 'URL'. If NULL, defaults to "https://api.nixtla.io/".}

\item{api_key}{The user's 'API' key. Get yours here: https://dashboard.nixtla.io/}
}
\value{
A message indicating the configuration status.
}
\description{
Set base 'ULR' and 'API' key in global environment
}
\examples{
\dontrun{
  nixtlar::nixtla_client_setup(
    base_url = "Base URL",
    api_key = "Your API key"
  )
}

}
