% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nixtla_set_api_key.R
\name{nixtla_set_api_key}
\alias{nixtla_set_api_key}
\title{Set 'API' key in global environment}
\usage{
nixtla_set_api_key(api_key)
}
\arguments{
\item{api_key}{The user's 'API' key. Get yours here: https://dashboard.nixtla.io/}
}
\value{
A message indicating the 'API' key has been set in the global environment.
}
\description{
This function will be deprecated in future versions. Please use \code{nixtla_client_setup} instead.
}
\examples{
\dontrun{
  nixtlar::nixtla_set_api_key("Your API key")
}

}
