% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nixtla_validate_api_key.R
\name{nixtla_validate_api_key}
\alias{nixtla_validate_api_key}
\title{Validate 'API' key}
\usage{
nixtla_validate_api_key()
}
\value{
TRUE if the API key is valid, FALSE otherwise.
}
\description{
Validate 'API' key
}
\examples{
\dontrun{
  nixtlar::nixtla_client_setup(api_key = "Your API key")
  nixtlar::nixtla_validate_api_key()
}

}
