% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nixtlaR-package.R
\docType{package}
\name{nixtlar-package}
\alias{nixtlar}
\alias{nixtlar-package}
\title{nixtlar: A Software Development Kit for 'Nixtla''s 'TimeGPT'}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

A Software Development Kit for working with 'Nixtla''s 'TimeGPT', a foundation model for time series forecasting. 'API' is an acronym for 'application programming interface'; this package allows users to interact with 'TimeGPT' via the 'API'. You can set and validate 'API' keys and generate forecasts via 'API' calls. It is compatible with 'tsibble' and base R. For more details visit \url{https://www.nixtla.io/docs}.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://nixtla.github.io/nixtlar/}
  \item \url{https://www.nixtla.io/docs}
  \item \url{https://github.com/Nixtla/nixtlar}
  \item Report bugs at \url{https://github.com/Nixtla/nixtlar/issues}
}

}
\author{
\strong{Maintainer}: Mariana Menchero \email{mariana@nixtla.io} (First author and maintainer)

Authors:
\itemize{
  \item Mariana Menchero \email{mariana@nixtla.io} (First author and maintainer)
}

Other contributors:
\itemize{
  \item Nixtla (Copyright held by 'Nixtla') [copyright holder]
}

}
\keyword{internal}
