% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nixtlaR-package.R
\docType{package}
\name{nixtlaR-package}
\alias{nixtlaR}
\alias{nixtlaR-package}
\title{nixtlaR: R SDK For TimeGPT}
\description{
An SDK to interface with Nixtla's TimeGPT from R. It provides functions for setting and validating tokens, and for generating forecasts with TimeGPT via API calls. Compatible with the tsibble package and base R.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://mmenchero.github.io/nixtlaR/}
}

}
\author{
\strong{Maintainer}: Mariana Menchero \email{mariana@nixtla.oi}

}
\keyword{internal}
