% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nixtlaR-package.R
\docType{package}
\name{nixtlaR-package}
\alias{nixtlaR}
\alias{nixtlaR-package}
\title{nixtlaR: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://mmenchero.github.io/nixtlaR/}
}

}
\author{
\strong{Maintainer}: First Last \email{first.last@example.com} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
\keyword{internal}
