% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nixtlaR-package.R
\docType{package}
\name{nixtlar-package}
\alias{nixtlar}
\alias{nixtlar-package}
\title{nixtlar: R SDK for Nixtla's TimeGPT}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

An SDK to interface with Nixtla's TimeGPT from R. It provides functions for setting and validating tokens and for generating forecasts with TimeGPT via API calls. Compatible with the tsibble package and base R.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://nixtla.github.io/nixtlar/}
}

}
\author{
\strong{Maintainer}: Mariana Menchero \email{mariana@nixtla.io} (First author and maintainer)

Other contributors:
\itemize{
  \item Nixtla (Copyright held by Nixtla) [copyright holder]
}

}
\keyword{internal}
