% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nixtlaR-package.R
\docType{package}
\name{nixtlar-package}
\alias{nixtlar}
\alias{nixtlar-package}
\title{nixtlar: An R SDK for 'Nixtla''s 'TimeGPT'}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

An R SDK for 'TimeGPT', a foundation model for time series forecasting developed by 'Nixtla'. Includes features for setting and validating 'API' keys and generating forecasts via 'API' calls. Compatible with 'tsibble' and base R. For more details, visit \url{https://docs.nixtla.io/}.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://nixtla.github.io/nixtlar/}
  \item \url{https://docs.nixtla.io/}
}

}
\author{
\strong{Maintainer}: Mariana Menchero \email{mariana@nixtla.io} (First author and maintainer)

Other contributors:
\itemize{
  \item Nixtla (Copyright held by Nixtla) [copyright holder]
}

}
\keyword{internal}
